pyproject.toml
[tool.poetry]
name = "pygame_sheetcut"
version = "1.0.0"
description = "A package for cutting sprite sheets using Pygame-CE"
authors = ["Maxwell Alexander Jackson max.the.scratcher@gmail.com"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pygame-ce = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"